Add a couple of unit tests
