--- src/running_mean_demo/__init__.py
@@ -1,2 +1,0 @@
-def main() -> None:
-    print("Hello from running-mean-demo!")
