+++ src/running_mean_demo/main.py
@@ -1,29 +1,11 @@
-import pydantic
+from . import elements
 import fastapi
 
 
 app = fastapi.FastAPI()
 
 
-class CurrentState(pydantic.BaseModel):
-    running_mean: float
-    count: int
-
-
-class NewValue(pydantic.BaseModel):
-    new_value: float
-
-
-def calculate_new_state(current_state: CurrentState, new_value: float) -> CurrentState:
-    """https://en.wikipedia.org/wiki/Moving_average#Cumulative_average"""
-    new_count = current_state.count + 1
-    new_mean = (
-        new_value + current_state.count * current_state.running_mean
-    ) / new_count
-    return CurrentState(running_mean=new_mean, count=new_count)
-
-
-current_state = CurrentState(running_mean=0.0, count=0)
+current_state = elements.CurrentState(running_mean=0.0, count=0)
 
 
 @app.get("/")
@@ -32,7 +14,9 @@ async def get_current():
 
 
 @app.post("/update/")
-async def update_running_mean(new_value: NewValue):  # -> CurrentState:
+async def update_running_mean(
+    new_value: elements.NewValue,
+):  # -> CurrentState:
     global current_state
-    current_state = calculate_new_state(current_state, new_value.new_value)
+    current_state = elements.calculate_new_state(current_state, new_value.new_value)
     return current_state
